+++ src/index.js
@@ -3,3 +3,4 @@
 export * from './core.js';
 export * from './elementals/accordion/index.js';
 export * from './elementals/disclosure/index.js';
+export * from './elementals/switch/index.js';